// src/shared/types/TaskUpdate.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskStatus } from "./TaskStatus";

export type TaskUpdate = { title: string | null, status: TaskStatus | null, owner: string | null, due: string | null, notes: string | null, };